pyproject.toml
[project]
name = "ethernity"
version = "0.1.0"
description = "Secure, offline-recoverable backup system with QR-based recovery documents"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.13"
keywords = ["backup", "recovery", "qr-code", "encryption", "offline", "paper-backup"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Archiving :: Backup",
    "Typing :: Typed",
]
dependencies = [
    "cbor2",
    "certifi",
    "mnemonic",
    "typer>=0.12.3",
    "rich>=13.7.0",
    "shellingham>=1.5.4",
    "questionary>=2.0.0",
    "platformdirs>=4.2.0",
    "fpdf2",
    "jinja2",
    "pycryptodome",
    "pillow",
    "pypdf",
    "segno",
    "zxing-cpp",
    "playwright>=1.57.0",
    "pyrage",
]

[project.scripts]
ethernity = "ethernity.cli:main"

[project.optional-dependencies]
build = ["pyinstaller>=6.10.0"]
dev = [
    "coverage>=7.13.1",
    "mypy>=1.11.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["ethernity", "ethernity.*"]

[tool.setuptools.package-data]
ethernity = ["templates/*/*.html.j2", "config/*.toml", "kit/*.html", "py.typed"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["src/ethernity"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
]
